requirements.txt
streamlit>=1.35
networkx>=3.0
pyvis>=0.3
pandas>=2.0
pyyaml>=6.0
plotly>=6.0.1

# Dokumenten-Loader
PyPDF2>=3.0
#Word docx
python-docx>=0.8.11
#Excel
openpyxl>=3.0.0
pdfplumber>=0.9.0

# Sonstiges
pydantic>=2.0